+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- Kubelet: `--cni-cache-dir` option, which defaults to `/var/lib/cni/cache`
 - [podman: 1.6.2](https://github.com/containers/libpod/releases/tag/v1.6.2)
 
 # 1.15.5 (17.10.2019)